single file int2bool direct encoding meta solver
Single file so it does not depend on a CPMpy version, and so that it
can be copied and used in side-projects without needing to be in master

Based on files Emilio sent me a month ago. I mostly refactored the top
part, not the actual encodings.

Only tested on sudoku, this can benefit from a lot more testing...