test comparisons between bool and int variables
from doc_rework, the test is literally from the new docs
(was failing for ortools, includes fix)

I had a look at adding it to test_constraints, but the generator to
which to add these cases is not clear to me...

if there is a way to simply test this with all solvers, without having
to add it to test_constraints, that would also be fine for me. Even
these as is just for ortools is already something.