SubCircuit(start) split into SubCircuitWithStart and SubCircuit
Split the SubCircuit global constraint with optional start_index into two constraints: SubCircuitWithStart & SubCircuit. SubCircuitWithStart is formulated on top of SubCircuit (which does not assume a start_index) and adds the condition for a provided start_index. This allows SubCircuitWithStart to still utilise solver's subcircuit, even when that solver does not support a start_index.