python_builtins: follow python signature
our 'min' and 'max' didn't allow multiple arguments, while Python's
does

in general we were hesitating whether our min/max/sum should mimic
Python's or NumPys. Now it was a mix in the middle... we were somewhat
like Python but did np.max() inside (same for sum, messing up the return
type); and we didnt yet get to implementing 'axis' like numpy either.

With this pull request, I propose that the least confusing thing for
people that do `from cpmpy import *` is to not break any existing
python code... that is, follow the Python signature of min/max/sum
(and fall back to that built-in when no decision variables are given).