Change "assumption variables" to "assumptions"
I believe this section is technically correct now. Since there can be non-Boolean assumptions (in the future), and because they should technically be Boolean literals, it is a bit unwise to call them assumption variables. Just calling them assumptions skirts the issue at least for this section, although there are other places in the documentation and code where they are called assumption variables.